requirements.txt
altgraph==0.17
appdirs==1.4.4
astroid==2.2.5
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.1
configparser==5.0.1
crayons==0.4.0
Cython==0.29.14
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
future==0.18.2
idna==2.8
importlib-metadata==3.7.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
mouse==0.7.0
MouseInfo==0.1.3
pefile==2019.4.18
Pillow==7.0.0
pipenv==2020.11.15
PyAutoGUI==0.9.52
PyAutoIt==0.4
pycodestyle==2.5.0
PyGetWindow==0.0.9
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pylint==2.3.1
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.1
python-anticaptcha==0.4.2
PyTweening==1.0.3
pywin32==225
pywin32-ctypes==0.2.0
requests==2.22.0
selenium==3.141.0
six==1.12.0
tesseract==0.1.3
typed-ast==1.4.0
typing-extensions==3.7.4.3
urllib3==1.25.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
webdriver-manager==3.2.2
Werkzeug==1.0.1
wrapt==1.11.2
zipp==3.4.1
